Rename brush config to sub brush
